Adding in handling for v2 iceberg tables
"show table extended" is used to gather information about the table in
order to make determinations about what type of table it is.  This sql
statement is not supported in V2 iceberg tables and the exception
generated when execute makes this very clear.  When this exception
happens rather than allowing the error to happen; another query is
performed to "show tables" instead which does not include extended
information.  When this returns a "describe extended" for the table is
executed to get extended information about the table.