Adding if/else for jc's changes where I can.
Removing hard coded paths and allowing for branches based on iceberg or
not.  I've tried if/else blocks where I can but some of the code needs
specific test cases because I'm unsure they are actually exercised by
the jaffle_shop dbt runs I'm doing and the unit/integration tests in
the project.